# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Satpol PP",
        "description": "Dokumentasi ini berisi semua endpoint API yang tersedia untuk proyek Satpol PP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/albums": {
            "get": {
                "tags": [
                    "Album"
                ],
                "summary": "Menampilkan semua album dengan paginasi",
                "operationId": "getAlbums",
                "responses": {
                    "200": {
                        "description": "Operasi berhasil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Album"
                ],
                "summary": "Membuat album baru",
                "operationId": "storeAlbum",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "cover": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/albums/{id}": {
            "get": {
                "tags": [
                    "Album"
                ],
                "summary": "Menampilkan detail album beserta foto-fotonya",
                "operationId": "getAlbumById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Album"
                ],
                "summary": "Memperbarui album (gunakan POST dengan _method=PUT)",
                "operationId": "updateAlbum",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "_method"
                                ],
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "judul": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "cover": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diperbarui"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Album"
                ],
                "summary": "Menghapus album dan semua fotonya",
                "operationId": "deleteAlbum",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}